requirement.txt
Flask
Pillow
qrcode[pil]
requests